+++ src/os2datascanner/projects/admin/adminapp/urls.py
@@ -20,7 +20,8 @@ from django.http import HttpResponse
 from django.views.i18n import JavaScriptCatalog
 from django.views.generic.base import TemplateView
 from os2datascanner import __version__, __commit__, __tag__, __branch__
-from os2datascanner.projects.admin.adminapp.views.analysis_views import AnalysisPageView
+from os2datascanner.projects.admin.adminapp.views.analysis_views import (AnalysisPageView,
+                                                                         AnalysisJobRunView)
 
 from os2datascanner.projects.admin import settings
 
@@ -95,8 +96,9 @@ urlpatterns = [
         template_name="openapi.yaml", content_type="application/yaml"),
         name="json-api"),
     re_path(r'^api/(?P<path>.*)$', JSONAPIView.as_view(), name="json-api"),
-    re_path(r'^analysis/$',  AnalysisPageView.as_view(), name='analysis'),
-
+    re_path(r'^analysis/overview/$',  AnalysisPageView.as_view(), name='analysis'),
+    re_path(r'^analysis/overview/(?P<pk>\d+)/run/$',
+            AnalysisJobRunView.as_view(), name='run-analysis-job'),
     # App URLs
     re_path(r'^status/$', StatusOverview.as_view(), name='status'),
     re_path(r'^status-completed/$', StatusCompleted.as_view(), name='status-completed'),