+++ src/os2datascanner/projects/static/js/charts/analysis/analysisPie-ChartJS3.js
@@ -8,30 +8,61 @@ function displayAsPercentage(value, ctx) {
   
   
 function displayLabelUnit(index, chartObject) {
-let dataset = chartObject.datasets[0];
-let labelName = chartObject.labels[index];
-if (dataset.name === "nfiles"){
-    return labelName.concat(": ", formatNumber(dataset.data[index]).toString(), " ", gettext("files"));
-}
-else if (dataset.name === "storage"){
-    return `${labelName}: ${bytesToSize(dataset.data[index])}`;
-}
-else {
-    return `${labelName}: ${formatNumber(dataset.data[index])}`;
-}
+    let dataset = chartObject.datasets[0];
+    let labelName = chartObject.labels[index];
+    const val = dataset.data[index];
+    if (dataset.name === "nfiles"){
+        if (val ===1){
+            return labelName.concat(": ", formatNumber(val).toString(), " ", gettext("file"));
+        }
+        else {
+            return labelName.concat(": ", formatNumber(val).toString(), " ", gettext("files"));
+        }
+    }
+    else if (dataset.name === "storage"){
+        return `${labelName}: ${bytesToSize(val)}`;
+    }
+    else {
+        return `${labelName}: ${formatNumber(val)}`;
+    }
 }
 
 function bytesToSize(bytes) {
-const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB'];
-if (bytes === 0) {return 'n/a';}
-const i = parseInt(Math.floor(Math.log(bytes) / Math.log(1024)));
-if (i === 0) {return `${bytes} ${sizes[i]}`;}
-return `${(bytes / Math.pow(1024, i)).toFixed(1)} ${sizes[i]}`;
+    const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB'];
+    if (bytes === 0) {return 'n/a';}
+    const i = parseInt(Math.floor(Math.log(bytes) / Math.log(1024)));
+    if (i === 0) {return `${bytes} ${sizes[i]}`;}
+    return `${(bytes / Math.pow(1024, i)).toFixed(1)} ${sizes[i]}`;
 }
 
 function formatNumber(x) {
-return x.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+    return x.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
 }
+
+function unorderedListLegend(data, colors, chart) {
+    let text = [`<ul id="${chart.id}" class="pie_legend_list">`];
+  
+    const totalValue = data.data.reduce((sum, obj) => sum + (obj || 0)); // Get the total of all non-zero data points
+  
+    for (i = 0; i < data.data.length; i++) {
+      const value = data.data[i];
+      const label = data.labels[i].charAt(0).toUpperCase() + data.labels[i].slice(1);
+      const backgroundColor = colors[i];
+  
+      const percentage = value ? ((value / totalValue) * 100).toFixed(2) + "%" : "0%"; // Calculate percentage based on totalValue
+  
+      if (value > 0) {
+        text.push(`<li><span class="bullet" style="color:${backgroundColor}">&#8226;</span>`);
+        if (label) {
+            text.push(`<span class="legend-txt">${label}</span>`,
+                      `<span class="data-label">${percentage}</span>`);
+        }
+        text.push('</li>');
+      }
+    }
+    text.push('</ul>');
+    return text.join("");
+  }
 
 function createPie(data, htmlElements, colors){ // jshint ignore:line
     let [ctx, pieCanvasID] = htmlElements;
@@ -45,9 +76,10 @@ function createPie(data, htmlElements, colors){ // jshint ignore:line
             datasets: [{
                 data: data.data,
                 backgroundColor: colors.slice(0, data.labels.length),
+                hoverBackgroundColor: colors.slice(0, data.labels.length),
                 borderColor: colors.slice(0, data.labels.length),
-                borderAlign: "center",
-                name: data.name
+                name: data.name,
+                hoverBorderWidth:0
             }],
         },
         plugins: [
@@ -107,7 +139,7 @@ function createPie(data, htmlElements, colors){ // jshint ignore:line
                 align: "center"
                 },
             tooltip: {
-                backgroundColor: 'white',
+                backgroundColor: 'rgba(255, 255, 255, 0.8)',
                 borderWidth: 1,
                 titleAlign: 'center',
                 bodyAlign: 'center',
@@ -142,6 +174,7 @@ function createPie(data, htmlElements, colors){ // jshint ignore:line
         maintainAspectRatio: false,
         }
     });
+    document.getElementById(ctx.id + "_legend").innerHTML = unorderedListLegend(data, colors, pie);
     return pie;
 }